Add PG classification to metadata
This adds support for classification to episode metadata.

Originally I wanted to add support for
**resource.images.classificationicons.colour** but Estuary does not
support PG classification icons, so the second best thing it to add it
to the metadata.

Apparently VTM GO only tags PG12 and PG16. There is no explicit
AL, PG6 or PG9 like in the Netherlands.